moved wallets state into WalletOptions.tsx
